# omod/src/main/webapp/resources/swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "REST API for Call flows Module.",
    "version" : "v1",
    "title" : "Call flows Module REST API"
  },
  "basePath" : "/openmrs/ws",
...
  "paths" : {
...
    "/callflows/flows" : {
      "get" : {
        "tags" : [ "REST API to manage call flows" ],
        "summary" : "Search for call flow",
        "description" : "Search for call flow",
        "operationId" : "searchFlowsByName",
        "parameters" : [ {
          "name" : "lookup",
          "in" : "query",
          "description" : "Lookup By Name",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "term",
          "in" : "query",
          "description" : "A search term that is interpreted by the concerned lookup function",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully searches a call flow",
            "schema" : {
              "$ref" : "#/definitions/SearchResponse"
            }
          }
        }
      },
      "post" : {
        "tags" : [ "REST API to manage call flows" ],
        "summary" : "Create a call flow",
        "description" : "Create a call flow if no duplicate (by name) exists",
        "operationId" : "createFlow",
        "parameters" : [ {
          "in" : "body",
          "name" : "callFlowRequest",
          "description" : "Call flow request with name, description and raw fields",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CallFlowRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully created a call flow",
            "schema" : {
              "$ref" : "#/definitions/CallFlowResponse"
            }
          },
          "400" : {
            "description" : "Call flow name is not an alphanumeric"
          },
          "409" : {
            "description" : "Call flow with the same name already exists"
          }
        }
      }
    },
    "/callflows/flows/{id}" : {
      "put" : {
        "tags" : [ "REST API to manage call flows" ],
        "summary" : "Update a call flow",
        "description" : "Update a call flow",
        "operationId" : "updateFlow",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Call flow id to update",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        }, {
          "in" : "body",
          "name" : "callFlowRequest",
          "description" : "Call flow request with name, description and raw fields",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/CallFlowRequest"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated a call flow",
            "schema" : {
              "$ref" : "#/definitions/CallFlowResponse"
            }
          },
          "400" : {
            "description" : "Call flow name is not an alphanumeric"
          },
          "409" : {
            "description" : "Call flow with the same name already exists"
          }
        }
      },
      "delete" : {
        "tags" : [ "REST API to manage call flows" ],
        "summary" : "Delete a call flow",
        "description" : "Delete a call flow",
        "operationId" : "deleteFlow",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "Call flow id to delete",
          "required" : true,
          "type" : "integer",
          "format" : "int64"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully deleted a call flow"
          },
          "400" : {
            "description" : "Unsuccessful to delete a call flow"
          }
        }
      }
    },
...
    "/callflows/status/{callId}" : {
      "get" : {
        "tags" : [ "REST API to update the call status periodically" ],
        "summary" : "Update the current call status",
        "description" : "Update the current call status",
        "operationId" : "handleStatus",
        "parameters" : [ {
          "name" : "callId",
          "in" : "path",
          "description" : "CallId to lookup by",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "params",
          "in" : "query",
          "description" : "Call status parameters",
          "required" : true,
          "type" : "object"
        } ],
        "responses" : {
          "200" : {
            "description" : "Successfully updated the current call status",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
  },
  "definitions" : {
    "CallFlowRequest" : {
      "type" : "object",
      "properties" : {
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "raw" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        }
      }
    },
    "CallFlowResponse" : {
      "type" : "object",
      "properties" : {
        "id" : {
          "type" : "integer",
          "format" : "int32"
        },
        "name" : {
          "type" : "string"
        },
        "description" : {
          "type" : "string"
        },
        "status" : {
          "type" : "string"
        },
        "raw" : {
          "type" : "string"
        }
      }
    },
...
    "SearchResponse" : {
      "type" : "object",
      "properties" : {
        "results" : {
          "type" : "array",
          "items" : {
            "type" : "object"
          }
        },
        "more" : {
          "type" : "boolean",
          "default" : false
        }
      }
    }
  }
}